requirements.txt
aiohttp==3.4.4
astroid==2.0.4
async-timeout==3.0.1
asyncpg==0.18.2
attrs==18.2.0
cffi==1.11.5
chardet==3.0.4
discord.py==1.1.1
emoji==0.5.2
idna==2.7
idna-ssl==1.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.5.0
pycparser==2.19
pylint==2.1.1
PyNaCl==1.2.1
regex==2019.5.25
six==1.11.0
typed-ast==1.1.0
websockets==6.0
wrapt==1.10.11
yarl==1.2.6
